backend/requirements.txt
# Framework Web
fastapi>=0.94.0,<0.95.0
uvicorn>=0.21.0,<0.22.0
pydantic>=1.10.7,<2.0.0
python-multipart>=0.0.6,<0.1.0

# Autenticação
python-jose>=3.3.0,<3.4.0
passlib>=1.7.4,<1.8.0
bcrypt>=4.0.1,<4.1.0

# Banco de Dados
sqlalchemy>=2.0.7,<2.1.0
alembic>=1.10.2,<1.11.0
psycopg2-binary>=2.9.5,<2.10.0  # Para PostgreSQL

# Conectores de Câmeras e Dispositivos
aiohttp>=3.8.4,<3.9.0  # Para requisições HTTP assíncronas
onvif-zeep>=0.2.12,<0.3.0  # Cliente ONVIF
netifaces>=0.11.0,<0.12.0  # Para obter informações de rede
ipaddress>=1.0.23,<1.1.0  # Para manipulação de endereços IP

# Processamento de Imagens
opencv-python-headless>=4.7.0.72,<4.8.0  # OpenCV sem dependências de GUI
numpy>=1.24.2,<1.25.0 
pillow>=9.4.0,<9.5.0  # Para manipulação de imagens

# Utilitários
python-dotenv>=1.0.0,<1.1.0
email-validator>=1.3.1,<1.4.0
httpx>=0.23.3,<0.24.0  # Cliente HTTP para testes

# Dependências para detecção avançada
motor>=3.1.2
aiofiles>=23.1.0
pymongo>=4.3.3
pytest>=7.3.1
pytest-asyncio>=0.21.0
ultralytics>=8.0.0  # YOLOv8 